# bag_vetscan/make_openapi_server/analyzer_app/openapi_server/openapi/openapi.yaml
openapi: 3.0.1
info:
  description: "The definition of the software interface between analyzers and the\
    \ HUB.  The interface will be implemented as a RESTful interface with all server\
    \ endpoints hosted on the Analyzer.\n\nThe following requirements will be met\
    \ by all interfaces: \n1. All data passed back from server shall be in JSON format.\n\
    2. All query parameters and JSON data properties shall be named using snake case\
    \ and be all lower case.\n4. All data types that describe a measurement value\
    \ shall end with an underscore followed by the unit of that physical value.  i.e.\
    \ motor_current_ma."
  title: Analyzer and HUB API
  version: "1.0"
servers:
- url: /
paths:
...
  /prompts/notification_ack:
    post:
      description: Hub is informing the analyzer a notification was acknowledged by
        the operator in response to a websocket message named notification on the
        prompts channel.
      operationId: prompts_notification_ack_post
      requestBody:
        $ref: '#/components/requestBodies/inline_object_3'
        content:
          application/json:
            schema:
              properties:
                correlation_id:
                  description: A unique ID that can be used to correlate messages
                    being sent and received
                  type: string
              required:
              - correlation_id
              type: object
        required: true
      responses:
        "200":
          description: Hub received message successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_400'
          description: This response is sent to a request that violates the predefined
            request schema
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_400'
          description: This response is sent to any request that the analyzer is unable
            to do at the time
      tags:
      - Prompts Channel
      x-openapi-router-controller: openapi_server.controllers.prompts_channel_controller
  /prompts/option_chosen:
    post:
      description: Hub is informing the analyzer of an option that was made by the
        operator in response to a websocket message named choose_option on the prompts
        channel.
      operationId: prompts_option_chosen_post
      requestBody:
        $ref: '#/components/requestBodies/inline_object_2'
        content:
          application/json:
            schema:
              properties:
                correlation_id:
                  description: A unique ID that can be used to correlate messages
                    being sent and received
                  type: string
                option_chosen:
                  description: The option chosen by the operator
                  example: "Yes"
                  type: string
              required:
              - correlation_id
              - option_chosen
              type: object
        required: true
      responses:
        "200":
          description: Hub received chosen option successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_400'
          description: This response is sent to a request that violates the predefined
            request schema
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_400'
          description: This response is sent to any request that the analyzer is unable
            to do at the time
      tags:
      - Prompts Channel
      x-openapi-router-controller: openapi_server.controllers.prompts_channel_controller
  /prompts/qr_scanned:
    post:
      description: Hub is informing the analyzer of a QR scan attempt in response
        to a websocket message named scan_qr on the prompts channel.
      operationId: prompts_qr_scanned_post
      requestBody:
        $ref: '#/components/requestBodies/inline_object_4'
        content:
          application/json:
            schema:
              properties:
                correlation_id:
                  description: A unique ID that can be used to correlate messages
                    being sent and received
                  type: string
                qr_contents:
                  description: The contents of the QR that was scanned.  If scanning
                    failed or was cancelled by operator, this will be an empty string
                  example: '{"exp":"2022-05-27","name":"c","lot":"088222","data":"12.34"}'
                  type: string
              required:
              - correlation_id
              - qr_contents
              type: object
        required: true
      responses:
        "200":
          description: Hub received message successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_400'
          description: This response is sent to a request that violates the predefined
            request schema
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_400'
          description: This response is sent to any request that the analyzer is unable
            to do at the time
      tags:
      - Prompts Channel
      x-openapi-router-controller: openapi_server.controllers.prompts_channel_controller
...
components:
...
  requestBodies:
...
    inline_object_4:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/inline_object_4'
      required: true
    inline_object_3:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/inline_object_3'
      required: true
    inline_object_2:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/inline_object_2'
      required: true
...
  schemas:
...
    inline_response_400:
      properties:
        detail:
          description: A description of the problem with the request
          title: detail
          type: string
        status:
          description: The number of the HTTP error code
          title: status
          type: number
        title:
          description: The title of the error indicated by the status
          title: title
          type: string
        type:
          description: Additional error information if available describing the type
            of error
          title: type
          type: string
      required:
      - detail
      - status
      - title
      - type
      title: inline_response_400
      type: object
...
    inline_object_2:
      properties:
        correlation_id:
          description: A unique ID that can be used to correlate messages being sent
            and received
          type: string
        option_chosen:
          description: The option chosen by the operator
          example: "Yes"
          type: string
      required:
      - correlation_id
      - option_chosen
      type: object
    inline_object_3:
      properties:
        correlation_id:
          description: A unique ID that can be used to correlate messages being sent
            and received
          type: string
      required:
      - correlation_id
      type: object
    inline_object_4:
      properties:
        correlation_id:
          description: A unique ID that can be used to correlate messages being sent
            and received
          type: string
        qr_contents:
          description: The contents of the QR that was scanned.  If scanning failed
            or was cancelled by operator, this will be an empty string
          example: '{"exp":"2022-05-27","name":"c","lot":"088222","data":"12.34"}'
          type: string
      required:
      - correlation_id
      - qr_contents
      type: object